requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.3.0
scikit-learn>=1.0.0
xgboost>=1.5.0
umap-learn>=0.5.0
scipy>=1.7.0